Feat: 동영상 정보 조회

+++ server/src/main/kotlin/com/bashplus/server/users/repository/UsersRepository.kt
@@ -13,7 +13,7 @@ import java.util.*
 open interface UsersRepository : JpaRepository<Users, String> {
     public fun findByIdAndType(id: String, type: String): Optional<Users>
 
-    public fun findByUid(uid: Int): Optional<Users>
+    public fun findByUid(uid: Long): Optional<Users>
 
     @Transactional
     @Modifying
